Use MathJax for rendering LaTeX
MathJax supports a great subset of LaTeX than the default rST setting of
using CSS.

This does not require any change on the part of the user. `math`
directives will simply now render LaTeX in a way that MathJax can
render.

One downside is that Encyclopedia now requires an Internet connection to
render math. Without, the raw LateX markup will be rendered instead.